requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
